Improved launcher and user feedback.
The launcher now uses JewelCli to parse program arguments.

While running, the applications keep track of statistics (currently
only the duration). These are printed to standard out after a
successful run.

I also found a cool testing library that allows system exit codes to
be tested, so there's now not a single line of code that can't be
tested!