Implemented the copy and delete actions.
The ordering of actions is not yet correct. It is fine for copying and
converting, but all delete actions should actually be done in reverse
order, otherwise the system will first delete a directory, and then the
files in it. That won't work.

The solution is clear: use a SortedSet<Action> instead of a List, and
order actions by putting deletions at the end, in reverse
lexicographical order.