Implemented ordering of actions, as well as logging.
The actions to be performed are now ordered using an ActionComparator,
which makes sure that actions are grouped together, and that all actions
in a single group are ordered correctly.

Shared code from the various Action classes is now in AbstractAction.

Added a SkipAction for files that are skipped. Reason: to make sure they
are logged.

Speaking of logging: each Action logs itself with the global
java.util.logging logger. The logging subsystem is configured to log to
the console cleanly.