Made copying of static files more efficient.
The actions for source content and static content are now merged in a
single set of actions. This merge ensures that files are not first
deleted (because they're not in the source content) and then copied
back (because there are in the target content).

Quite a refactoring was needed to make it work, but the result is
indeed better, since the new ActionSet encapsulates tricky logic that
was first handled implicitly by the Magisto class (or not at all).