Implemented support to "force overwrite" files.
Using the `-f` command line argument, Magisto will not process files
that it would otherwise skip because the source file hasn't changed.

I added this flag because it can be useful sometimes. Consider:

- Create a file `image.jpg` in the source tree.
- Create a file `image.jpg` in the static tree (`.static`)
- Run Magisto
- Delete the file `image.jpg` from the source tree.

The file `image.jpg` will not be replaced with the one in the static
tree, since Magisto has no way of knowing that its source has changed.

This is an edge case. A more typical case is when you've changed the
custom page template and you want all files to be regenerated.

Anyhow, using `-f` does the trick.