Added a custom `link` directive to the FreeMarker configuration.
The `link` directive makes it possible to link to files relative to the
source of the publication from the custom template.

Let's say to want to link to stylesheet from your template:

    <link rel="stylesheet" type="text/css" href="static/style.css"/>

This link will only resolve correctly for Markdown files in the root
of the source tree. Putting a `/` in front of the `static` is not a
good solution, because then the output MUST be published at root level
on a web server.

By applying the `link` directive:

    <link rel="stylesheet" type="text/css"
         href="<@link path="/static/style.css"/>"/>

You always end up with a working link.