Added automatic detection of changes to the custom template.
If the custom template has changed, all Markdown files are automatically
converted again, even if they haven't changed.

I added this because I was messing around with my custom template a lot
and had to use the `-f` option a lot, which also resulted in
unnecessary copying of all static content and non-Markdown files in
the source root.