Replaced maven-shade-plugin with maven-assembly-plugin.
The distribution is now built as a separate ZIP that can be unpacked and
installed. I did this in preparation of the Maven plugin.

The `/bin/sh` script works on my OS X box, and I'm pretty sure it'll
also work on Linux.

The Windows script I couldn't test, so it might not work.