Added the Git repository changelog to the page model.
Each page template now receives the changelog: the last - at most 30 -
changes in the repository. A template can use this to generate a nice
overview of the changes.

I do this now in the template for a specific site, but only if the
name of the file being processed is called `changelog.md`. The result
is a dedicated page for this overview.