Replaced the global changelog with a per-file changelog.
I clearly didn't have my thinking cap on when I implemented the global
changelog. It's very nice that you can make a single page that is
amended by the page template to write the complete log, but when is
that changelog being updated? If the source file doesn't change, then
it won't. And it typically isn't. A bit of Catch-22.

Also, it already felt a bit ugly, to pass global information to each
individual page.

The best way to solve it was to remove it completely. But I did
replace it with a per-page history. So it's still possible to show a
list of changes. Just not a global one ;-)