Made the JGitHistory cache the Git repository data.
This is a big performance improvement, since the default template
goes through all commits twice - once of a table, once for a detailed
list - and through the last commit many changes.

It was being fetched from the repository every time.

Now the data is only fetched once, and only if accessed by the template.