Implement a custom Markdown parser to fix broken anchors for headers.
The default implementation of the Parser can generate anchors for
headers. It does that by appending all the text in the header.
Unfortunately it skips any text within code markers (``).

For example:

    # This is `code`

...becomes:

    <a name="this-is-"></a>

I've fixed that by overriding the collectChildren method from the
default Parser and changing the logic a tiny bit. Now it generates:

    <a name="this-is-code"></a>

...which is exactly what I would expect.